+++ lib/sentinel/helpers/view_helper.ex
@@ -1,0 +1,5 @@
+defmodule Sentinel.ViewHelper do
+  def user_view do
+    Application.get_env(:sentinel, :user_view)
+  end
+end