+++ test/support/user_view_test_helper.exs
@@ -1,0 +1,23 @@
+defmodule Sentinel.UserView do
+  use Phoenix.View, root: "web/templates"
+  import Phoenix.Controller, only: [view_module: 1]
+
+  def render("index.json", %{users: users}) do
+    render_many(users, Sentinel.UserView, "user.json")
+  end
+
+  def render("show.json", %{user: user}) do
+    render_one(user, Sentinel.UserView, "user.json")
+  end
+
+  def render("user.json", %{user: user}) do
+    %{id: user.id,
+      email: user.email,
+      role: user.role,
+      hashed_password: user.hashed_password,
+      hashed_confirmation_token: user.hashed_confirmation_token,
+      confirmed_at: user.confirmed_at,
+      hashed_password_reset_token: user.hashed_password_reset_token,
+      unconfirmed_email: user.unconfirmed_email}
+  end
+end