+++ test/test_helper.exs
@@ -5,4 +5,5 @@ Application.put_env(:phoenix, :format_encoders, [json: Poison])
 
 Code.require_file "test/support/ecto_helper.exs"
 Code.require_file "test/support/router_helper.exs"
+Code.require_file "test/support/user_view_test_helper.exs"
 Code.require_file "test/support/forge.exs"