+++ CHANGELOG.md
@@ -1,1 +1,8 @@
 # Changelog
+## 0.0.5
+Makes `guardian_db` dependency optional
+
+Updates controllers to use 201 status codes where appropriate
+
+Updates Accounts Controller to render according to the `user_view`
+config